refactor: extract begin function to not stick to journey
